requirements.txt
# Core MCP Server Framework
elevenlabs-mcp

# FastAPI and related dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# YouTube audio download
yt-dlp>=2023.12.30

# Local audio transcription
openai-whisper>=20231117

# Google Gemini LLM (free tier)
google-generativeai>=0.3.0

# Local Text-to-Speech
TTS>=0.22.0

# HTTP requests for Figma API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Additional dependencies for audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Image processing for design tool
Pillow>=10.0.0

# JSON handling and utilities
aiofiles>=23.0.0
